ui(refactor): split 'superdesk-dictionaries' module.
