+++ scripts/superdesk-dictionaries/controllers/index.js
@@ -1,0 +1,2 @@
+export { DictionaryConfigController } from './DictionaryConfigController';
+export { DictionaryEditController } from './DictionaryEditController';
