+++ scripts/superdesk-dictionaries/directives/DictionaryConfigModal.js
@@ -1,0 +1,7 @@
+export function DictionaryConfigModal() {
+    return {
+        controller: 'DictionaryEdit',
+        require: '^sdDictionaryConfig',
+        templateUrl: 'scripts/superdesk-dictionaries/views/dictionary-config-modal.html'
+    };
+}
