+++ scripts/superdesk-dictionaries/directives/index.js
@@ -1,0 +1,2 @@
+export { DictionaryConfigModal } from './DictionaryConfigModal';
+export { FileUpload } from './FileUpload';
