+++ scripts/superdesk-dictionaries/services/index.js
@@ -1,0 +1,1 @@
+export { DictionaryService } from './DictionaryService';
